Move the lookup parsing!!
Oh man, this feels good.
